feat: cmd dispatcher; relocate handlers
